rozdelenie do tried
